Robustness + DRY: safe-close helper, dedup bit helpers, lazy hot-path logging, fix connect log
- Extract ModbusController._safe_close(); replace 5 copies of the close-on-error
  boilerplate.
- Move set_bit/get_bit_bool to helpers (were defined verbatim in both
  solis_binary_sensor and solis_select_entity); re-import for compatibility.
- Poll loop now logs unexpected errors at warning with exc_info instead of
  swallowing them at debug.
- Fix the inverted/misleading serial-number log on connect (was INFO every
  connect, logging 'from device: None'); now a single debug line.
- Convert the per-register/per-block/per-sensor hot-path f-string debug logs to
  lazy %s form so they don't build strings when debug is off.
